Move helper files into src
